Make date of post file name congruent with date in post metadata.
This ought not to be necessary. The date metadata in the file is the date
of publication and rules. The file name date is merely for convenience.
I've found that relative-url does not work with the files whose date metadata
differs from the date filename. So here. Now they'll match.

There are a few, the potable water tank files, to go because I want to test
the links to them and not orphan them.